+++ src/main/java/realtech/db/entity/G5WriteYbGallery02.java
@@ -87,7 +87,7 @@ public class G5WriteYbGallery02 implements Serializable {
 	private String wrFacebookUser;
 
 	@Column(name="wr_file")
-	private byte wrFile;
+	private int wrFile;
 
 	@Column(name="wr_good")
 	private int wrGood;
@@ -102,7 +102,7 @@ public class G5WriteYbGallery02 implements Serializable {
 	private String wrIp;
 
 	@Column(name="wr_is_comment")
-	private byte wrIsComment;
+	private int wrIsComment;
 
 	@Column(name="wr_last")
 	private String wrLast;
@@ -306,7 +306,7 @@ public class G5WriteYbGallery02 implements Serializable {
 		this.wrFacebookUser = wrFacebookUser;
 	}
 
-	public byte getWrFile() {
+	public int getWrFile() {
 		return this.wrFile;
 	}
 
@@ -346,7 +346,7 @@ public class G5WriteYbGallery02 implements Serializable {
 		this.wrIp = wrIp;
 	}
 
-	public byte getWrIsComment() {
+	public int getWrIsComment() {
 		return this.wrIsComment;
 	}
